Backend/requirements.txt
Flask==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
numpy==1.26.0